+++ giford/action/reshape.py
@@ -2,8 +2,8 @@ import enum
 import numpy as np
 from skimage import transform
 
-from giford.image_actions.image_action import ChainImageAction
-from giford.image_actions.translate import Translate
+from giford.action.abstract_frame_action import ChainFrameAction
+from giford.action.translate import Translate
 from giford.frame_batch import FrameBatch
 from giford.raw_data import RawDataFrame
 
@@ -14,7 +14,7 @@ class ReshapeMethod(enum.Enum):
     DOWNSCALE = 3
 
 
-class Reshape(ChainImageAction):
+class Reshape(ChainFrameAction):
     """
     turns 120x120 into 60x60 or 240x240
     """
