+++ giford/action/translate.py
@@ -2,14 +2,14 @@ import ffmpeg
 from skimage import transform
 import numpy as np
 
-from giford.image_actions.image_action import ChainImageAction
+from giford.action.abstract_frame_action import ChainFrameAction
 from giford.frame_batch import FrameBatch
 from giford.raw_data import RawDataFrame
 
 from giford.virtual_path import Movement
 
 
-class Translate(ChainImageAction):
+class Translate(ChainFrameAction):
     def __init__(self):
         super().__init__()
