+++ tests/test_swirls.py
@@ -3,7 +3,7 @@ import pytest
 
 from giford.frame_batch import FrameBatch
 from giford.image import SingleImage
-from giford.image_actions.swirl import BasicSwirl, VariableSwirl, VaryingVariableSwirl
+from giford.action.swirl import BasicSwirl, VariableSwirl, VaryingVariableSwirl
 from tests.util import (
     BASELINE_DIRECTORY,
     TEST_INPUT_ORANGE_IMAGE_FILEPATH,
